Create LoginController class for interface
