+++ src/main/java/interface/controller/LoginController.java
@@ -1,5 +1,7 @@
 package app.view;
 
+import javafx.scene.layout.HBox;
+
 public class LoginView {
   
 }
