--- src/main/java/interface/view/UI.java
@@ -1,28 +1,0 @@
-package ui;
-
-import javafx.application.Application;
-import javafx.scene.Scene;
-import javafx.scene.control.Label;
-import javafx.scene.layout.StackPane;
-import javafx.stage.Stage;
-
-public class UI extends Application {
-
-    @Override
-    public void start(Stage primaryStage) throws Exception {
-      //primaryStage.setTitle("My First JavaFX App");
-
-      //primaryStage.show();
-        String javaVersion = System.getProperty("java.version");
-        String javafxVersion = System.getProperty("javafx.version");
-        Label l = new Label("Hello, JavaFX " + javafxVersion + ", running on Java " + javaVersion + ".");
-        Scene scene = new Scene(new StackPane(l), 640, 480);
-        primaryStage.setScene(scene);
-        primaryStage.show();
-    }
-
-    public static void main(String[] args) {
-        launch();
-    }
-
-}
